Extract ImageGallery component
