+++ src/components/ImageGallery/ImageGallery.module.scss
@@ -1,0 +1,73 @@
+.gallery {
+    width: 70%;
+    position: relative;
+
+    .counter {
+        width: 60px;
+        height: 60px;
+        background-color: white;
+        border-radius: 30px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        position: absolute;
+        top: calc(50% - 30px);
+        right: 30px;
+    }
+
+    .navigationLeft, .navigationRight {
+        cursor: pointer;
+        width: 96px;
+        height: 96px;
+        position: absolute;
+        top: calc(50% - 48px);
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        z-index: 100;
+
+        & img {
+            width: 100%;
+            height: 100%;
+        }
+    }
+
+    .navigationLeft {
+        left: -60px;
+    }
+
+    .navigationRight {
+        right: -60px
+    }
+
+    :global(.image-gallery-slides) {
+        line-height: 0;
+        overflow: hidden;
+        position: relative;
+        white-space: nowrap;
+        text-align: center;
+    }
+
+    :global(.image-gallery-slide) {
+        left: 0;
+        position: absolute;
+        top: 0;
+        width: 100%;
+
+        & :global(.image-gallery-image) {
+            width: 100%;
+            object-fit: contain;
+        }
+    }
+
+    :global(.image-gallery-slide.center) {
+        position: relative;
+    }
+    
+    .image {
+        & img {
+            max-width: 100%;
+            max-height: 100%;
+        }
+    }
+}
